+++ uamp-client/CHANGELOG.md
@@ -1,4 +1,5 @@
 # CHANGELOG
 
+- Add progress bar dragging
 - Fix volume overwrite on startup
 - Fix album hover effect